github/vulnerable_requirements.txt
Flask==1.1.1
Django==1.11.29
requests==2.19.0
urllib3==1.24.1
PyYAML==3.13
Jinja2==2.10.1
Werkzeug==0.16.1
MarkupSafe==1.1.1
itsdangerous==1.1.0
click==7.0
blinker==1.4
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
certifi==2019.6.16
chardet==3.0.4
idna==2.8
numpy==1.16.6
pandas==0.24.2
scikit-learn==0.20.4
matplotlib==3.1.1
seaborn==0.9.0
plotly==4.1.1
dash==1.0.2
dash-core-components==1.0.0
dash-html-components==1.0.0
dash-table==4.0.0
dash-renderer==1.0.0
dash-bootstrap-components==0.7.2
dash-daq==0.1.0
dash-cytoscape==0.1.0
dash-sunburst==0.0.1
dash-canvas==0.1.0
dash-bio==0.1.0
dash-ag-grid==0.1.0
dash-mantine-components==0.1.0
dash-extensions==0.0.1
dash-iconify==0.1.0
dash-leaflet==0.1.0
dash-plotly-editor==0.1.0
dash-slicer==0.1.0
dash-snapshots==0.1.0
dash-testing==0.1.0
dash-uploader==0.1.0
dash-cytoscape==0.1.0
dash-sunburst==0.0.1
dash-canvas==0.1.0
dash-bio==0.1.0
dash-ag-grid==0.1.0
dash-mantine-components==0.1.0
dash-extensions==0.0.1
dash-iconify==0.1.0
dash-leaflet==0.1.0
dash-plotly-editor==0.1.0
dash-slicer==0.1.0
dash-snapshots==0.1.0
dash-testing==0.1.0
dash-uploader==0.1.0
